Implemented time command, other minor changes
Implemented the time command where the user can input a time in the format 'YYYY-MM-DD HH'. The command can also be used with an argument. For this a function to convert a string to unixtime has also been made

changed the command functions to all take the entire program_state as an argument, and instead handling which values to change in the function itself. this lowers the amount of arguments for each function, and also standardizes the command function to all use the same argument (which is pretty cool)

changed the reset command to reset the entire program state instead of just the screen